issue-30 move assertSingleJobIsDistributedBetweenWorkers to StubbedMultiJob, delete VerificationUtils
